finally a full setup
