+++ src/extension.ts
@@ -1,4 +1,5 @@
 import * as vscode from "vscode";
+import { ACAIResourcesViewProvider } from "./providers/ACAIResourcesViewProvider/ACAIResourcesViewProvider";
 
 export function activate(context: vscode.ExtensionContext) {
   const provider = new ACAIResourcesViewProvider(context.extensionUri);
@@ -9,56 +10,47 @@ export function activate(context: vscode.ExtensionContext) {
       provider
     )
   );
+
+  // Register commands
+  context.subscriptions.push(
+    vscode.commands.registerCommand(
+      "acai-resources-tool.refreshResources",
+      () => {
+        vscode.window.showInformationMessage("Refreshing ACAI resources...");
+        // Add logic to refresh resources
+      }
+    )
+  );
+
+  context.subscriptions.push(
+    vscode.commands.registerCommand("acai-resources-tool.addResource", () => {
+      vscode.window
+        .showInputBox({
+          prompt: "Enter the name of the new resource",
+          placeHolder: "New Resource",
+        })
+        .then((resourceName) => {
+          if (resourceName) {
+            vscode.window.showInformationMessage(
+              `Adding new resource: ${resourceName}`
+            );
+            // Add logic to add a new resource
+          }
+        });
+    })
+  );
+
+  // Add event listener for configuration changes
+  context.subscriptions.push(
+    vscode.workspace.onDidChangeConfiguration((e) => {
+      if (e.affectsConfiguration("acai-resources-tool")) {
+        vscode.window.showInformationMessage(
+          "ACAI Resources Tool configuration changed"
+        );
+        // Add logic to handle configuration changes
+      }
+    })
+  );
 }
 
-class ACAIResourcesViewProvider implements vscode.WebviewViewProvider {
-  public static readonly viewType = "acai-resources-sidebar";
-
-  constructor(private readonly _extensionUri: vscode.Uri) {}
-
-  public resolveWebviewView(
-    webviewView: vscode.WebviewView,
-    context: vscode.WebviewViewResolveContext,
-    _token: vscode.CancellationToken
-  ) {
-    webviewView.webview.options = {
-      enableScripts: true,
-      localResourceRoots: [this._extensionUri],
-    };
-
-    webviewView.webview.html = this._getHtmlForWebview(webviewView.webview);
-  }
-
-  private _getHtmlForWebview(webview: vscode.Webview) {
-    const scriptUri = webview.asWebviewUri(
-      vscode.Uri.joinPath(
-        this._extensionUri,
-        "dist",
-        "ACAIResourceView",
-        "index.js"
-      )
-    );
-    const styleUri = webview.asWebviewUri(
-      vscode.Uri.joinPath(
-        this._extensionUri,
-        "dist",
-        "ACAIResourceView",
-        "index.css"
-      )
-    );
-
-    return `<!DOCTYPE html>
-            <html lang="en">
-            <head>
-                <meta charset="UTF-8">
-                <meta name="viewport" content="width=device-width, initial-scale=1.0">
-                <link href="${styleUri}" rel="stylesheet">
-                <title>ACAI Resources</title>
-            </head>
-            <body>
-                <div id="root"></div>
-                <script src="${scriptUri}"></script>
-            </body>
-            </html>`;
-  }
-}
+export function deactivate() {}
