+++ src/providers/ACAIResourcesViewProvider/ACAIResourcesViewProvider.ts
@@ -1,0 +1,62 @@
+import * as vscode from "vscode";
+
+export class ACAIResourcesViewProvider implements vscode.WebviewViewProvider {
+  public static readonly viewType = "acai-resources-sidebar";
+
+  constructor(private readonly _extensionUri: vscode.Uri) {}
+
+  public resolveWebviewView(
+    webviewView: vscode.WebviewView,
+    context: vscode.WebviewViewResolveContext,
+    _token: vscode.CancellationToken
+  ) {
+    webviewView.webview.options = {
+      enableScripts: true,
+      localResourceRoots: [this._extensionUri],
+    };
+
+    webviewView.webview.html = this._getHtmlForWebview(webviewView.webview);
+
+    // Add message listener
+    webviewView.webview.onDidReceiveMessage((message) => {
+      switch (message.command) {
+        case "alert":
+          vscode.window.showInformationMessage(message.text);
+          return;
+      }
+    });
+  }
+
+  private _getHtmlForWebview(webview: vscode.Webview) {
+    const scriptUri = webview.asWebviewUri(
+      vscode.Uri.joinPath(
+        this._extensionUri,
+        "dist",
+        "ACAIResourceView",
+        "index.js"
+      )
+    );
+    const styleUri = webview.asWebviewUri(
+      vscode.Uri.joinPath(
+        this._extensionUri,
+        "dist",
+        "ACAIResourceView",
+        "index.css"
+      )
+    );
+
+    return `<!DOCTYPE html>
+            <html lang="en">
+            <head>
+                <meta charset="UTF-8">
+                <meta name="viewport" content="width=device-width, initial-scale=1.0">
+                <link href="${styleUri}" rel="stylesheet">
+                <title>ACAI Resources</title>
+            </head>
+            <body>
+                <div id="root"></div>
+                <script src="${scriptUri}"></script>
+            </body>
+            </html>`;
+  }
+}
